Extend facebox to allow modal dialogs.
We needed the ability to force a user to complete a dialog, without the option
to cancel it. Facebox didn't support this, so I decided to add that as a new
setting 'modal'. While I was in there, I changed it to also prefetch its images
from a Rails-idiomatic load path.